% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_engine_path.R
\name{stata_engine_path}
\alias{stata_engine_path}
\alias{have_stata_engine_path}
\title{Stata Engine Path}
\usage{
stata_engine_path()

have_stata_engine_path()
}
\value{
Path to stata engine
}
\description{
Determines the path to the Stata Engine
}
\examples{
stata_engine_path()
have_stata_engine_path()
}
