% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_engine_path_worker.R
\name{stata_engine_path_worker}
\alias{stata_engine_path_worker}
\title{Stata Engine Path Background Worker}
\usage{
stata_engine_path_worker(SE = FALSE, version = 14)
}
\arguments{
\item{SE}{is the Stata you have Special Edition?}

\item{version}{Which version of Stata do you have (only relevant on Windows)}
}
\value{
Path to stata engine
}
\description{
Helps Determines the path to the Stata Engine
}
\examples{
stata_engine_path_worker()
}
