% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_knitr_opts.R
\name{stata_knitr_opts}
\alias{stata_knitr_opts}
\title{Options in knitr for Stata}
\usage{
stata_knitr_opts(engine.path = stata_engine_path(), comment = "", ...)
}
\arguments{
\item{engine.path}{path for stata engine}

\item{comment}{comment for output, passed to \code{\link[knitr]{opts_chunk}}}

\item{...}{additional arguments passed to passed to \code{\link[knitr]{opts_chunk}}}
}
\value{
Output of \code{\link[knitr]{opts_chunk}}
}
\description{
Uses \code{\link[knitr]{opts_chunk}} to set options for knitting the markdown.
}
\examples{
stata_knitr_opts()
}
